Correctly handle the time zone of recorded times
The recorded times are interpreted as being in the Budapest time zone
(which is what ZakStunts uses) and then converted to UTC before any
further processing.

Time zone handling matters here because of daylight saving time. For
instance, the PTB window of ZCT284 was 599 hours long rather than the
usual 600 because the window included the start of DST in Budapest.